docs/public/course/llm-agent/code-examples/requirements.txt
# LLM Agent 开发指南 - 依赖包列表

# 核心 LLM 库
openai>=1.50.0
anthropic>=0.34.0

# 数学和科学计算
numpy>=1.21.0
scipy>=1.7.0
sympy>=1.9.0
matplotlib>=3.5.0

# 数据处理和验证
pandas>=1.3.0
pydantic>=1.8.0

# 模板引擎
jinja2>=3.0.0

# 环境变量管理
python-dotenv>=0.19.0

# HTTP 请求
requests>=2.25.0
aiohttp>=3.8.0

# 异步支持
asyncio-mqtt>=0.11.0

# 可视化增强
plotly>=5.0.0
seaborn>=0.11.0

# 开发和测试工具
pytest>=6.0.0
pytest-asyncio>=0.18.0
black>=21.0.0
flake8>=3.9.0

# 文档生成
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Jupyter 支持（可选）
jupyter>=1.0.0
ipykernel>=6.0.0

# 其他实用工具
tqdm>=4.60.0
colorama>=0.4.4
rich>=10.0.0
